# swagger.json
{
    "swagger":"2.0",
    "info":{
       "title":"Classified-ads API",
       "description":"Check the API, any help or bug fix is really appreciated! <a href=\"github.com/bacloud22/classified-ads/issues\">Open an issue or follow with a pull request</a>",
       "version":"1.0.0"
    },
...
    "paths":{
...
       "/listings/gwoogl":{
          "post":{
             "parameters":[
                {
                   "name":"body",
                   "in":"body",
                   "schema":{
                      "$schema":"http://json-schema.org/draft-07/schema#",
                      "type":"object",
                      "properties":{
                         "title_desc":{
                            "type":"string",
                            "minLength":3,
                            "maxLength":100
                         },
                         "exact":{
                            "type":"boolean",
                            "default":false
                         },
                         "div_q":{
                            "type":"string",
                            "minLength":3,
                            "maxLength":40
                         },
                         "since":{
                            "type":"string",
                            "format":"date"
                         },
                         "section":{
                            "type":"string",
                            "enum":[
                               "donations",
                               "skills"
                            ]
                         }
                      },
                      "required":[
                         "title_desc",
                         "section"
                      ]
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"Default Response"
                }
             }
          }
       },
       "/listings/geolocation":{
          "post":{
             "parameters":[
                {
                   "name":"body",
                   "in":"body",
                   "schema":{
                      "$schema":"http://json-schema.org/draft-07/schema#",
                      "type":"object",
                      "properties":{
                         "lat":{
                            "type":"number",
                            "maximum":90,
                            "minimum":-90
                         },
                         "lng":{
                            "type":"number",
                            "maximum":180,
                            "minimum":-180
                         },
                         "section":{
                            "type":"string",
                            "enum":[
                               "donations",
                               "skills"
                            ]
                         }
                      },
                      "required":[
                         "lat",
                         "lng",
                         "section"
                      ]
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"Default Response"
                }
             }
          }
       },
...
    },
    "host":"localhost:3000",
    "schemes":[
       "http"
    ],
    "consumes":[
       "application/json"
    ],
    "produces":[
       "application/json"
    ],
...
 }